fix(CatalogService): refactor)
