--- packages/CatalogService/hekp-for-pnpm.md
@@ -1,1 +1,0 @@
-pnpm add cors -E -S --filter @telicent-oss/catalogservice
