+++ packages/CatalogService/src/__tests__/utils/formatDataAsArray.ts
