+++ packages/CatalogService/src/__tests__/utils/makeStatic.ts
