+++ packages/CatalogService/src/apiFactory/operations/utils/instanceFromResourceFactory.ts
@@ -1,5 +1,5 @@
 import { HumanError } from "../../../utils/HumanError";
-import { CatalogService } from "../../../index";
+import { CatalogService } from "../../../../index";
 import { ResourceType, ResourceSchema, DCATResourceSchema } from "./common";
 import { printJSON } from "./printJSON";
 import { tryInstantiate } from "./tryInstantiate";
