+++ packages/CatalogService/src/apiFactory/operations/utils/printJSON.ts
