+++ packages/CatalogService/src/apiFactory/operations/utils/transformRdfToTree.ts
