+++ packages/CatalogService/src/apiFactory/operations/utils/tryCatch.ts
